+++ apps/client/windows/flutter/generated_plugins.cmake
@@ -10,6 +10,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   flutter_timezone
   flutter_tts
   pasteboard
+  share_plus
   speech_to_text_windows
   url_launcher_windows
 )